Version 0.5 - Conversational Graph
Implemented basic functionality that allows the user to engage in a CLI conversation with GPT that can in-turn query and write to a Graph Database hosted on Neo4j.

This is very basic functionality in that there is no real conversation history (yet) for follow up questions. You can simply ask it to perform lookups or create nodes, it turns requests into Cypher queries, and passes them in. Subsequent releases will add functionality.